Get rid of hack based on old DevUI to exchange information between ORM and Flyway extensions
